% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcfR.R
\name{io_vcfR}
\alias{io_vcfR}
\alias{write.vcf}
\title{io_vcfR}
\usage{
write.vcf(x, file = "", mask = FALSE, APPEND = FALSE)
}
\arguments{
\item{x}{vcfR object}

\item{file}{outfile name}

\item{mask}{use mask?}

\item{APPEND}{append to existing file?}
}
\description{
input and output of VCF data.
}
\details{
input and output of VCF data.
}

