% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKitty.R
\docType{package}
\name{myKitty}
\alias{myKitty}
\alias{myKitty-package}
\title{myKitty: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{myKitty functions}{

The foo functions ...
}

