% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_gz_write}
\alias{test_gz_write}
\title{test_gz_write}
\usage{
test_gz_write(filename = "myFile.gz", mask = 0L)
}
\arguments{
\item{filename}{name for created file}

\item{mask}{integer}
}
\description{
test of writing gz file
}
\details{
test of writing gz file
}

